Move model interfaces to contracts directory
