+++ app/Contracts/MentionAble.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Contracts;
 
+use App\Models\User;
 use Illuminate\Support\Collection;
 
 interface MentionAble
