+++ app/Contracts/ReplyAble.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Contracts;
 
 use Illuminate\Database\Eloquent\Relations\MorphMany;
