+++ app/Contracts/SubscriptionAble.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Contracts;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 interface SubscriptionAble
