+++ app/Http/Requests/CreateReplyRequest.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Requests;
 
-use App\Models\ReplyAble;
+use App\Contracts\ReplyAble;
 use App\Models\Thread;
 use App\Models\User;
 use App\Rules\HttpImageRule;
