+++ app/Models/Thread.php
@@ -12,6 +12,9 @@ use App\Concerns\HasUuid;
 use App\Concerns\PreparesSearch;
 use App\Concerns\ProvidesSubscriptions;
 use App\Concerns\ReceivesReplies;
+use App\Contracts\MentionAble;
+use App\Contracts\ReplyAble;
+use App\Contracts\SubscriptionAble;
 use App\Exceptions\CouldNotMarkReplyAsSolution;
 use Exception;
 use Illuminate\Contracts\Pagination\Paginator;
